Refactor: extract scraper and filters modules
