+++ index.js
@@ -5,6 +5,8 @@ const sqlite3 = require('sqlite3').verbose();
 const path = require('path');
 const { OpenAI } = require('openai');
 const { parseOpenAIResponse, getFirstSentence } = require('./lib/extractParties');
+const { runFilters } = require("./lib/filters");
+const { scrapeSource } = require("./lib/scraper");
 
 const openai = new OpenAI({ apiKey: process.env.OPENAI_API_KEY });
 
@@ -336,87 +338,7 @@ app.delete('/filters/:id', (req, res) => {
   });
 });
 
-async function runFilters(articleIds, logs) {
-  if (!articleIds.length) return;
-  const filters = await new Promise((resolve, reject) => {
-    db.all('SELECT * FROM filters WHERE active = 1', [], (err, rows) => {
-      if (err) return reject(err);
-      resolve(rows);
-    });
-  });
 
-  if (!filters.length) {
-    logs && logs.push('No active filters to run');
-    return;
-  }
-
-  for (const id of articleIds) {
-    const article = await new Promise((resolve, reject) => {
-      db.get('SELECT title, description FROM articles WHERE id = ?', [id], (err, row) => {
-        if (err) return reject(err);
-        resolve(row);
-      });
-    });
-    if (!article) continue;
-
-    for (const filter of filters) {
-      if (filter.type === 'keyword') {
-        const keywords = (filter.value || '')
-          .split(',')
-          .map(k => k.trim().toLowerCase())
-          .filter(Boolean);
-        const text = `${article.title || ''} ${article.description || ''}`.toLowerCase();
-        const matched = keywords.some(kw => text.includes(kw));
-        if (matched) {
-          await new Promise((resolve, reject) => {
-            db.run(
-              'INSERT INTO article_filter_matches (article_id, filter_id) VALUES (?, ?)',
-              [id, filter.id],
-              err => (err ? reject(err) : resolve())
-            );
-          });
-        }
-      } else if (filter.type === 'embedding') {
-        // TODO: implement semantic filtering using embeddings
-      }
-    }
-  }
-  logs && logs.push(`Ran ${filters.length} filters on ${articleIds.length} articles`);
-}
-
-async function scrapeSource(source) {
-  const response = await axios.get(source.base_url);
-  const $ = cheerio.load(response.data);
-
-  const articles = [];
-  $(source.article_selector).each((i, el) => {
-    const container = $(el);
-    let time = '';
-    if (source.time_selector) {
-      time = container.find(source.time_selector).text().trim();
-      container.find(source.time_selector).remove();
-    }
-    const title = container.find(source.title_selector).text().trim();
-    const description = source.description_selector
-      ? container.find(source.description_selector).text().trim()
-      : '';
-    let link = source.link_selector
-      ? container.find(source.link_selector).attr('href') || ''
-      : '';
-    if (link && !link.startsWith('http')) {
-      try {
-        link = new URL(link, source.base_url).href;
-      } catch (e) {}
-    }
-    const image = source.image_selector
-      ? container.find(source.image_selector).attr('src') || null
-      : null;
-
-    articles.push({ title, description, time, link, image });
-  });
-
-  return articles;
-}
 
 // Scrape endpoint
 app.get('/scrape', async (req, res) => {
@@ -464,7 +386,7 @@ app.get('/scrape', async (req, res) => {
       insertedTotal += inserted;
       logs.push(`Inserted ${inserted} new articles from ${source.base_url}`);
       if (insertedIds.length) {
-        await runFilters(insertedIds, logs);
+        await runFilters(db, insertedIds, logs);
       }
       details.push({
         source_id: source.id,
@@ -501,7 +423,7 @@ app.get('/run-filters', async (req, res) => {
     });
     logs.push('Cleared previous filter matches');
 
-    await runFilters(ids, logs);
+    await runFilters(db, ids, logs);
     res.json({ processed: ids.length, logs });
   } catch (err) {
     console.error(err);
